claimclear_v2_healthcare (1)/claimclear_v2_healthcare (1)/claimclear_v3/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
groq==0.9.0
chromadb==0.5.3
sentence-transformers==3.0.1
pymupdf==1.24.5
pypdf==4.2.0
pytesseract==0.3.10
Pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
numpy==1.26.4
